+++ bufferkdtree/src/neighbors/brute/base.c
@@ -7,64 +7,52 @@
 #include "include/gpu_opencl.h"
 #include "include/global.h"
 
-// global parameters
-Parameters params_global;
-
-// timers
-DEFINE_TIMER(1);
-DEFINE_TIMER(2);
-DEFINE_TIMER(3);
-
 /* -------------------------------------------------------------------------------- 
  * Interface (extern): Initialize components
  * --------------------------------------------------------------------------------
  */
 void init_extern(int n_neighbors, int num_threads, int platform_id, \
-		int device_id, char *kernels_source_directory, int verbosity_level) {
+		int device_id, char *kernels_source_directory, int verbosity_level, \
+		BRUTE_PARAMETERS *params) {
 
-	START_TIMER(1);
+	set_default_parameters(params);
 
-	set_default_parameters(&params_global);
-	params_global.n_neighbors = n_neighbors;
-	params_global.num_threads = num_threads;
-	params_global.platform_id = platform_id;
-	params_global.device_id = device_id;
-	params_global.kernels_source_directory = kernels_source_directory;
-	params_global.verbosity_level = verbosity_level;
-	check_parameters(&params_global);
+	params->n_neighbors = n_neighbors;
+	params->num_threads = num_threads;
+	params->platform_id = platform_id;
+	params->device_id = device_id;
+	params->kernels_source_directory = kernels_source_directory;
+	params->verbosity_level = verbosity_level;
 
-	INIT();
+	check_parameters(params);
 
-	STOP_TIMER(1);
-
-	PRINT("Initializing time (extern): \t\t%2.10f\n", (FLOAT_TYPE) GET_TIME(1));
-
-}
-
-void print_parameters_extern(void) {
-	printf("=================================== Parameter Settings ===================================\n");
-	printf("Number of nearest neighbors (K_NN): %i\n", params_global.n_neighbors);
-	printf("Number of threads for CPU (num_threads): %i\n", params_global.num_threads);
-	printf("Level of verbosity (verbosity_level): %i\n", params_global.verbosity_level);
-	printf("Double precision? (USE_DOUBLE): %i\n", USE_DOUBLE);
-	printf("Using GPU device (USE_GPU): %i\n", USE_GPU);
-	printf("Floating point precision: %s\n", PLOT_PRECISION_INFOS());
-	printf("==========================================================================================\n");
 }
 
 /* --------------------------------------------------------------------------------
  * Interface (extern): fit model
  * -------------------------------------------------------------------------------- 
  */
-void fit_extern(FLOAT_TYPE *X, int nX, int dX) {
+void fit_extern(FLOAT_TYPE *X, int nX, int dX, BRUTE_RECORD *brute_record,
+		BRUTE_PARAMETERS *params) {
 
-	PRINT("Fitting model ...\n");
+	int i;
+	for (i = 0; i < 25; i++) {
+		INIT_MY_TIMER(brute_record->timers + i);
+	}
 
-	START_TIMER(2);
-	FIT(X, nX, dX, &params_global);
-	STOP_TIMER(2);
+	brute_record->dXtrain = dX;
+	brute_record->nXtrain = nX;
 
-	PRINT("Fitting time (extern): \t\t\t\t\t\t\t\t%2.10f\n", (FLOAT_TYPE) GET_TIME(2));
+	INIT(brute_record, params);
+
+	PRINT(params)("Fitting model ...\n");
+
+	START_MY_TIMER(brute_record->timers + 1);
+	FIT(X, nX, dX, brute_record, params);
+	STOP_MY_TIMER(brute_record->timers + 1);
+
+	PRINT(params)("Fitting time (extern): \t\t\t\t\t\t\t\t%2.10f\n", \
+			(FLOAT_TYPE) GET_MY_TIMER(brute_record->timers + 1));
 
 }
 
@@ -74,15 +62,17 @@ void fit_extern(FLOAT_TYPE *X, int nX, int dX) {
  */
 void neighbors_extern(FLOAT_TYPE *Xtest, int nXtest, int dXtest,
 		FLOAT_TYPE* distances, int ndistances, int ddistances,
-		int* indices, int nindices, int dindices) {
+		int* indices, int nindices, int dindices,
+		BRUTE_RECORD *brute_record, BRUTE_PARAMETERS *params) {
 
-	PRINT("Computing nearest neighbors ...\n");
+	PRINT(params)("Computing nearest neighbors ...\n");
 
-	START_TIMER(3);
-	NEIGHBORS(Xtest, nXtest, dXtest, distances, indices, &params_global);
-	STOP_TIMER(3);
+	START_MY_TIMER(brute_record->timers + 2);
+	NEIGHBORS(Xtest, nXtest, dXtest, distances, indices, brute_record, params);
+	STOP_MY_TIMER(brute_record->timers + 2);
 
-	PRINT("Total computation time (extern): \t\t\t\t\t\t%2.10f\n", (FLOAT_TYPE) GET_TIME(3));
+	PRINT(params)("Total computation time (extern): \t\t\t\t\t\t%2.10f\n", \
+			(FLOAT_TYPE) GET_MY_TIMER(brute_record->timers + 2));
 
 }
 
@@ -90,9 +80,18 @@ void neighbors_extern(FLOAT_TYPE *Xtest, int nXtest, int dXtest,
  * Frees some resources (e.g., on the GPU)
  * --------------------------------------------------------------------------------
  */
-void free_resources_extern(void) {
+void free_resources_extern(BRUTE_RECORD *brute_record, BRUTE_PARAMETERS *params) {
 
-	FREE_RESOURCES();
+	FREE_RESOURCES(brute_record, params);
 
 }
 
+/* --------------------------------------------------------------------------------
+ * Checks if platform and device are valid
+ * --------------------------------------------------------------------------------
+ */
+int extern_check_platform_device(int platform_id, int device_id){
+
+	return get_device_infos(platform_id, device_id, NULL);
+
+}