+++ examples/benchmark.py
@@ -16,7 +16,7 @@ import generate
 from bufferkdtree.neighbors import NearestNeighbors
 
 # platform dependent parameters
-# (adapt if needed!)
+# (might have to be adapted!)
 plat_dev_ids = {0:[0,1,2,3]}
 n_jobs = 8
 
@@ -30,21 +30,23 @@ n_neighbors = 10
 results = {}
 
 print("Parsing data ...")
-Xtrain, Ytrain, Xtest = generate.get_data_set(NUM_TRAIN=200000, NUM_TEST=10000000)
-print "-------------------------------- DATA --------------------------------"
-print "Number of training patterns:\t", Xtrain.shape[0]
-print "Number of test patterns:\t", Xtest.shape[0]
-print "Dimensionality of patterns:\t", Xtrain.shape[1]
-print "----------------------------------------------------------------------"
+Xtrain, Ytrain, Xtest = generate.get_data_set(NUM_TRAIN=2000000, NUM_TEST=10000000)
+print("-------------------------------- DATA --------------------------------")
+print("Number of training patterns:\t %i" % Xtrain.shape[0])
+print("Number of test patterns:\t %i" % Xtest.shape[0])
+print("Dimensionality of patterns:\t%i" % Xtrain.shape[1])
+print("----------------------------------------------------------------------")
 
 def run_algorithm(n_test, algorithm="buffer_kd_tree"):
 
-    print("\n\nRunning %s for n_test=%i ..." % (algorithm, n_test))
-
+    print("----------------------------------------------------------------------")
+    print("\n\nRunning %s for n_test=%i ...\n" % (algorithm, n_test))
+    print("----------------------------------------------------------------------")
     opt_tree_depth = None
     Xtest_local = Xtest[:n_test, :]
 
     if algorithm in ["buffer_kd_tree", "kd_tree"]:
+
         # the different tree depths that shall
         # be tested for this data set
         if algorithm == "buffer_kd_tree":
@@ -62,19 +64,24 @@ def run_algorithm(n_test, algorithm="buffer_kd_tree"):
                                                     target="test", tree_depths=tree_depths)
         print("Optimal tree depth found: %i " % opt_tree_depth)
 
+        # instantiate final model
         nbrs = NearestNeighbors(n_neighbors=n_neighbors, \
                                 algorithm=algorithm, \
                                 n_jobs=n_jobs, \
                                 tree_depth=opt_tree_depth, \
                                 plat_dev_ids=plat_dev_ids, \
                                 verbose=verbose)
+
     elif algorithm == "brute":
+
         nbrs = NearestNeighbors(n_neighbors=n_neighbors, \
                                 algorithm=algorithm, \
                                 n_jobs=n_jobs, \
                                 plat_dev_ids=plat_dev_ids, \
                                 verbose=verbose)
+
     else:
+
         raise Exception("Unkown algorithm ...")
         
     # train model
@@ -84,9 +91,10 @@ def run_algorithm(n_test, algorithm="buffer_kd_tree"):
     train_time = (end_time-start_time)
     print("Fitting time: %f" % train_time)
 
-    # test model
+    # apply model (testing phase)
     start_time = time.time()
     dists, inds = nbrs.kneighbors(Xtest_local)
+    print "dists=", dists
     end_time = time.time()
     test_time = (end_time-start_time)
     print("Testing time: %f" % test_time)