+++ docs/changes.rst
@@ -4,8 +4,9 @@ Changes
 =======
 
 Release 1.1.0 (November 2015)
-----------------------------
+-----------------------------
 * Fixed wrong parameter assignment in 'kneighbors' method of both neighbors/kd_tree/base.py and neighbors/buffer_kdtree/base.py
+* Added Multi-GPU support to brute-force approach (for benchmark purposes)
 * Added benchmark example
 
 Release 1.0.2 (September 2015)